+++ rbx_xml/CHANGELOG.md
@@ -1,9 +1,14 @@
 # rbx_xml Changelog
 
 ## Unreleased
+
+# 3.0.0 (2026-07-01)
+* Upgraded to `rbx_reflection` v3.0.0
 * Added support for one-to-many property migrations. ([#612])
+* Added support for converting `SharedString` and `NetAssetRef` values to other values when necessary. ([#634])
 
 [#612]: https://github.com/rojo-rbx/rbx-dom/pull/612
+[#634]: https://github.com/rojo-rbx/rbx-dom/pull/634
 
 ## 2.0.1 (2025-11-27)
 * Upgrade rbx-dom dependencies.